+++ packages/web-components/src/app.tsx
@@ -106,6 +106,21 @@ const App = () => {
           'start-countdown': true,
         });
         break;
+      case 'enhanced-smartselfie-capture':
+        componentElement = createElement('enhanced-smartselfie-capture', props);
+        break;
+      case 'enhanced-smart-selfie-consent':
+        componentElement = createElement(
+          'enhanced-smart-selfie-consent',
+          props,
+        );
+        break;
+      case 'enhanced-smart-selfie-submission':
+        componentElement = createElement(
+          'enhanced-smart-selfie-submission',
+          props,
+        );
+        break;
       case 'smart-camera-web':
       default:
         componentElement = createElement('smart-camera-web', props);
@@ -181,6 +196,7 @@ const App = () => {
       key: `${component}-${themeColor}`,
       'show-navigation': true,
       'theme-color': themeColor,
+      'auto-capture': true,
     };
 
     switch (component) {